Fix upgrade auto-restart with 0 retries
Summary: If you do `hh_client --retries 0`, the first time we'll kill
the server, the second time we'll start it back up, and the third time
will work (if the server is up by that point). We should start the
server in the same pass we kill it in -- that shouldn't count as a
retry.

Reviewed By: @elgenie

Differential Revision: D1507218