Covariance fixes
Summary: This diff updates the typechecker to support covariance for the following
interfaces: Traversable, Container, Iterable, Iterator, ConstCollection,
KeyedTraversable, KeyedContainer, Indexish, KeyedIterable, KeyedIterator.

This diff also two bugs with covariance. First, it fixes the logic for
checking if array<..> is a subtype of Indexish<..>, Traversable<..>,
KeyedTraversable<..>, Container<..>, or KeyedContainer<..>. Second, it
fixes a bug in the covariance logic that was erroneously allowing certain
subtype relationships (for example, it was allowing a Pair to be passed to
a function expecting an ImmMap).

Reviewed By: @jwatzman

Differential Revision: D1511248