Implemented xdebug feature_get and feature_set commands
Summary: This diff implements the xdebug remote debugging commands feature_get
and feature_set. Like the name implies, these allow you to get and set
the features supported by remote debugging.

These are the first commands intended to be sent/received when starting
a debugging session, so these are important. Right now the results are
identical to xdebug (which is a lie because not all features are implemented).

Each command is also considered a feature, so this required me to implement the
stubs for each remote debugging command, which is why this diff is so large.
Most of the code is boilerplate, the most of the added logic is in the feature
get and feature set commands.

In this diff I also move `XDebugServer::initDbgp` and `XDebugServer::deinitDbgp`
into `XDebugServer::attach` and `XDebugServer::detach`. This way commands can
assume `XDEBUG_GLOBAL(Server)` always exists. It was nonintuitive that it
didn't, so I felt this was important for maintenance purposes.

Reviewed By: @ptarjan

Differential Revision: D1502891