Imported xdebug xml variable exporting and implemented xdebug eval + condition breakpoints
Summary: This diff was primarily intended to implement eval and condition breakpoints.
To do so, however, you need to be able to output the resulting variable in
an xml format. So a large part of this diff is pulling in the xdebug variable
xml exporting functions.

The imported code has been heavily rewritten to support hhvm internals, pretty
much only the flow and the function names remain. Tthe code is really ugly,
primarily due to the xml api so I've put added a note to refactor once the
xdebug xml api has been refactored.

I've also only imported the xml related exporting functions. Importing the rest
of the file would take too much time, but it isn't overly difficult. It will need
to be done for xdebug var_dump.@paulbiss is working on importing some other
parts of the file.

This took a while to import, but after this the commands should all fall in
place.

Reviewed By: @ptarjan

Differential Revision: D1507246