Delete unused stat counters
Summary: All of these are unused. I suspect some of them were originally meant to
count something that we still care about (e.g. tracelet guard
executions?) but the callsites have bitrotted away.

I think we'll need automated testing if we care about a particular stat
enough to keep it. I'd like to check this diff in as-is, but it's also
partially a forcing function to get people to reinstate stat counters
that they actually care about.

Reviewed By: @edwinsmith

Differential Revision: D1515024