Handle exact substring matches within fuzzy search
Summary: This addresses what was mentioned here: https://www.prod.facebook.com/groups/devtools.feedback/permalink/720320464683238/

In any case where the letter expected in the exact match appeared beforehand somewhere in the string, we ignored the exact match case.  This handles that, so any time we find an exact substring match we'll score it properly.

Reviewed By: @joelbeales

Differential Revision: D1517792