Let vasm use all the registers
Summary: Expand the unreserved registers for vasm (when compiling a whole
IRUnit, but not when compiling snippets of code via Vauto) to include
all usable registers (everything except rbx,r12,rsp,rbp).

Add an "args" mask to all the call opcodes to indicate which registers
are used (as register arguments) to the call. vasm-xls marks these used
by the call; prior to the call, there are copies into these registers,
so they are live from the copy to the call.

Reviewed By: @swtaarrs

Differential Revision: D1527672