Assign vasm registers to all SSATmps
Summary: Flip x64 code generation over to virtual registers. If all uses and
defs of an SSATmp only access a whole cell, assign a "wide" location.
If a tmp is defined by DefConst, assign a cpool-backed virtual register.
If the type does not have a known DataType, assign a type register.
This intentionally ignores src/dest constraints, which will be phased
out.

Reviewed By: @swtaarrs

Differential Revision: D1530245