(HHVM) Don't malloc/free for pack
Summary: `pack` is used by Thrift, and we were also thinking of using it for the
TAO client as a cheap way to turn integers into strings. I noticed that
it uses `std::vector` even if the format is only a handful characters
long. 32, because why not. Who uses format strings longer than 32?

Reviewed By: @jano, @markw65

Differential Revision: D1520273