Add a mechanism for building and testing a zend extension as a DSO.
Summary: There is a DSO, and a (currently failing) unit test.

This DSO is intended to be loaded by specifying in a .ini file something like:
  hhvm.dynamic_extensions[dso_test] = ./runtime/ext_zend_compat/dso_test/dso_test.so
and then tested by executing the php functions implemented by the dso_test zend extension.

The tests currently fail, and as such are parked in:
  hphp/test/zend/bad/ext/ext_dso_test

The tests fail due to one or more known logic bugs involving the handling of configuration
settings in zend DSOs in the file:
  hphp/runtime/ext_zend_compat/php-src/Zend/zend_ini.cpp
A fix for those bugs, as well as migrating the tests to hphp/test/slow/ext_dso_test,
will follow as soon as these changes are accepted.

This is the 2nd attempt at this work.  The 1st attempt:
  https://github.com/facebook/hhvm/pull/3399
  https://reviews.facebook.net/D21369
was aborted due to a poor choice of directory names/locations,
as well as a botched git rebase.  Paul Tarjan commented on the 1st attempt.

Closes https://github.com/facebook/hhvm/pull/3441

Reviewed By: @ptarjan

Differential Revision: D1492264

Pulled By: @sgolemon