Overhauled xdebug remote debugging testing
Summary: This diff does several things:
* Moved remote debugging tests into a single folder (eliminates the need for a
  bunch of the same ini files)
* Removed null characters as well as session-specific information from the
  output of the server before displaying. This means expectf files will now show
  up in phabricator, and most of the time out2expect.py is enough to generate
  a .expectf file
* Changed remote_FOO.test -> FOO.php.test
* FOO.test.in files can have comments now (1st character on line must be #)
* Removed script status output, unecessary and it made writing expectfs a pain
* FOO.php files are all the same now, just a single line. The testing file is
  determined automatically

This was high-pri on my list of cleanup as the old way was awful and not very
maintainable. I have most of the commands implemented and wanted this before
I wrote tests.

arc patch isn't happy with these binary files, but it's all just test motion and tests
are passing, so...

Reviewed By: @ptarjan

Differential Revision: D1510839