Fix bogus type assertions from nested arrays
Summary: In the new test case hhbbc correctly figured out that after
the first [], we had an array all of whose keys were intergers,
and all of whose values were array(55). For the second [] it knew
that the base was either null, or array(55). As a result, it gave up
on trying to update the known value of the array - but didn't discard
what it already knew.

When we iterated the array, it then thought the value had to be null, or
array(55). We eliminate the null by the if/continue, and so it asserted
that it was an array of one element (it doesn't bother to assert the types).

There are two fixes. The first was to improve the type of the base. If the
base of an array is known to be null or array(55), and we push somthing onto
it, we should promote it to empty|array(55), and then do the push (the existing
code converted it to null|empty|array(55)). The second was to drop any
specialized information from the base type when we don't know how to update it.

Reviewed By: @edwinsmith

Differential Revision: D1529980