Fix lto build
Summary: Even though we forced a fake call to const_load, and marked it
EXTERNALLY_VISIBLE, the lto build still optimized the call away
and then gold's --gc-sections option caused it to be removed.

I tried NEVER_INLINE, and it didn't help; apparently g++ notices that
its a no-op, and doesn't bother calling it (which I suppose is not the same
as inlining it).

This just forces a reference to its address, which makes it stick around.

Reviewed By: @edwinsmith

Differential Revision: D1571856