Initial vasm -> llvm backend
Summary: This can't emit any useful code yet, but I'm going to spend some time
working on vasm before I can make more progress so I'd rather get this in
now. It contains shell code to convert vasm to llir and emit machine code from
that, with support for punting if an unsupported operation is encountered. I
wrote up maybe-correct translations for a couple instructions and verified that
llvm is able to at least emit some code into a, and we're able to rewind the
frontier and then emit real code using vasm-x64.

Reviewed By: @jasone

Differential Revision: D1534862