Make RUSAGE_THREAD available to getrusage PHP function
Summary: The signature of the PHP function is getrusage( [ int $who = 0 ] ).
In both HHVM and PHP5, $who can be 1 to request RUSAGE_CHILDREN data. This
change allows $who to be 2 to request RUSAGE_THREAD data. RUSAGE_THREAD
measures resource usage for just the calling thread.

This allows HHVM PHP code to measure resource usage for the current request.
(In PHP5 this can be done with RUSAGE_SELF, by dint of the threading model.)
Closes https://github.com/facebook/hhvm/pull/3632

Reviewed By: @JoelMarcey

Differential Revision: D1530059