Support for HH_FIXME
Summary: It's sometimes difficult to fix all the typing errors at once.
This problem occurs when we try to convert a core library or when
we introduce a new check that reveals a lot of typing errors.

This diff adds support for a new construction: one can now add
/* HH_FIXME[ERROR_NUMBER] */ on the line that contains an error and
the error will be ignored by the type-checker.

We hope that this will help us to crowdsource the conversion effort.

Reviewed By: @jwatzman, @elgenie

Differential Revision: D1517560