Fix read of uninitialized stack in AGetL interp
Summary: If an interpreted AGetL ended up fataling or otherwise encountering an
exception, it would read an uninitialized stack slot during unwinding.
Make sure it's initialized to uninit-null instead. This will never be
seen by user code; either it will be overwritten by the right classref
or we'll unwind past it.

Reviewed By: @jdelong

Differential Revision: D1536360