Improved fix for issues with optional specialized array types in interp-minstr
Summary: Moves the changes to the handleBaseFooElem functions, and add some
documentation about what they are supposed to do now.  The bug was basically
that these functions were assuming that they didn't need to do anything about
inner-array-element cases, because the callers own that, and the callers were
assuming they only needed to continue array chains or otherwise deal with these
situations when the base type was a subtype of TArr.  But you'll notice in this
that no one was taking responsibility for handling the case where the base
couldBe(TArr), which needs to have some changes in case we're in the middle of
an array chain or operating directly on a local.  This makes handling that part
of the handleBaseFooElem functions.

Reviewed By: @markw65

Differential Revision: D1540665