Analyze/optimize pseudomains in hhbbc
Summary: We need to have a pass that determines which static properties can be
modified by the program (including public ones), which means we need to see
possible changes to static properties in pseudomains.  The easiest way to do
this is to analyze psuedomains like anything else.  (This also should help with
more clearly seeing the differences in the repo stats when hphpc's inferTypes
and postOptimize are off, since most of the bytecode changes are currently in
pseudomains.)

Reviewed By: @alexmalyshev

Differential Revision: D1545270