Allow vm-stack flushing in rpc threads
Summary: We disabled the flush in rpc threads, on the grounds that the
execution context lives across requests. However, the execution context
is never *required* to live across requests, and could be reset for a
number of reasons at the start of the next request.

So instead, we reset the request, and then flush. This also solves (at
least in practice) the issue of rpc requests keeping treadmill fromd
working.

Reviewed By: @bertmaher, @dariorussi

Differential Revision: D1546288