Optimize strtr
Summary: For an input string of length n, and a replacement array with
minlen and maxlen as the pattern string lengths, strtr() makes
O(n * (maxlen - minlen)) calls to Array::exists().

Since the input string is different each time, Array::exists() has
to compute the hash of the string every time.

This diff optimizes strtr() for the case when the replacement
array has less than 64 elements. It uses a 64 bit bitmask to
compute if a potential match exists, and makes the Array::exists()
calls if the possibility of a match is not eliminated.

Reviewed By: @markw65

Differential Revision: D1546956