Migrate MySQL configs to ::Bind
Summary: Doing this proactively now for a few sites in order for my bad ini node diff to work without a lot of test runner `.expect` changes.

Plus this is the right thing to do anyway.

Note: I removed the `IniSetting::Bind` line because that is now duplicative to the new `Config::Bind` line above it.

Reviewed By: @ptarjan

Differential Revision: D1547854