Move MemoryManager::refreshStatsImpl out of the header
Summary: Since MM is included by the world it's hard to iterate on this code
rapidly.  This function doesn't appear to be perf critical (and is too big for
inlining to make a huge difference anyways), so move it to the cpp file and
declare the boolean template specializations explicitly.

Reviewed By: @edwinsmith

Differential Revision: D1552525