Always assign Vloc register 0.
Summary: Currently we assign a const register to DefConst's results,
but if a non-DefConst instruction has a const result, we assign
0 registers. With this diff, Vloc.reg(0) always holds a valid
register. reg(1) is still optional; if a subtype of Gen has a
known DataType, or can be given a SIMD register, then it only
gets 1 register instead of 2.

After this lands we can start removing lots of code that checks
for InvalidReg (it will never happen), or refactor it to check
SSATmp::isConst() rather than the presence of a register.

Reviewed By: @ottoni

Differential Revision: D1554886