Fixed an assertion in the emitter for IncDec of $this
Summary: While processing $this++, the emitter would first push $this onto the
eval stack as a cell, then it would try to emit an IncDec bytecode instruction.
However, there exists no variant of IncDec which operates on the top value in
the eval stack. As it is always a syntax error to try to IncDec all other
cells, we kept the invariant that IncDec will assert if it encounters a cell on
the eval stack. We now check explicity for the $this++ case in visit().

Reviewed By: @jdelong

Differential Revision: D1545839