Don't allow split before phidef in the same block.
Summary: If a phidef caused another interval to be split and spilled,
the split position could be just before the phidef in the same
block, which is illegal. Fix the nearestSplitPosition to
be the start of the block, which forces copies to the edge.

Added an assert that checks we don't try to copy two different
intervals to the same physical register along the same edge.
(which fires before this fix).

Reviewed By: @alexmalyshev

Differential Revision: D1568347