Abort the server if a request gets stuck
Summary: If a thread gets stuck in C++ land (in extensions or HHVM runtime),
then it's not killed by the surprise mechanism.  Such requests end up
backing up the treadmill, increasing memory pressure, until HHVM is
eventually killed by the OOM killer (or dies when running out of low
memory).  This diff tries to detect such cases by checking the age of
the oldest request in the treadmill, and then aborting the server if
the request has been running for too long.  The thread running the
oldest request is identified and sent a SIGABRT, so that its backtrace
is output to the stacktrace file.

Reviewed By: @ptarjan

Differential Revision: D1544907