Move profCounter out of Func and into type-profile.cpp
Summary: type-profile.cpp handled marking them hot based on it anyway, and
knows when the data can't be used anymore to throw it away.  Doing this lets us
shuffle the fields slightly to go from 136 bytes (where I guess jemalloc will
allocate 160) to 128 (a size class).  Unfortunately for closures we're usually
going to waste 24 bytes now (or for funcs with more prologues than the fixed
number it could be similar).  We might want to consider putting that stuff
somewhere else so all Funcs can be the same size?  (Then we could pool allocate
them and not worry about mapping to a size class exactly.)

Reviewed By: @jasone

Differential Revision: D1560194