Add surprise checks to range()
Summary: range() can be made to allocate an unbounded amount of memory in a C++
loop, so we need surprise checks. Do it every 1000 steps (no scientific
justification for this number). The `unsigned char` overload of
ArrayUtil::Range doesn't need this because it can result in an array of
at most 256 elements.

Reviewed By: @edwinsmith

Differential Revision: D1552594