Prefix magic attributes with "__"
Summary: See linked task for full description, but here's the gist. In
the absence of real namespacing for user attributes, we should do the
same thing as PHP did for reserved functions: anything prefixed by `__`
might be magical in the runtime (in our case, in the typechecker).
Anything else is up for grabs by users of the language.

In that vein, prefix the handful of unprefixed magic attributes with
`__`. Maintain backwards compatibility support for the old names for a
little while.

Going forward, we should only make magic for user attributes prefixed
with `__`.

Namespacing support for user attributes might be a good idea, at which
point we can redo this, but it's complicated and no one is volunteering
to do it, so here we are.

Reviewed By: @int3

Differential Revision: D1559079