Revert "Custom path to the installed folder"
Summary: I'm torn on this so want to use this for discussion. Should the INSTALL_PATH namespace where we get `/etc/hhvm/php.ini` from? I'm assuming it was added for someone on those other systems, but I do know that on apple things are read from `/etc` without namespacing to `/usr/local` so I'm assuming the rest of that is wrong. Thoughts?

Fixes https://github.com/hhvm/packaging/issues/71

Reviewed By: @paulbiss

Differential Revision: D1560037