stop breaking ubuntu 10.04
Summary: A partial revert of https://github.com/facebook/hhvm/pull/3591 . I used `FindLibZip.cmake` as a template. I think for system defined libraries the old code was good enough, but if the thing is only installed in the CMAKE_INCLUDE_PATH then the macro@​radford was using won't work.

Reviewed By: @JoelMarcey

Differential Revision: D1560119