Added a rule to the simplifier for LdCtx to constant fold when inlined
Summary: When compiling LateBoundCls, we emit a LdCtx followed by a LdClsCtx to
get the Class that we're operating on. When LateBoundCls is inlined in such a
way where we statically know which Class the callee was invoked with, we can
constant fold the LdCtx and LdClsCtx, which then allows us to DCE the
DefInlineFP and other associated boilerplate for reifying an inline call frame.

Reviewed By: @jdelong

Differential Revision: D1552775