hh_ide.js - Reduce memory usage
Summary: We're starting to bleed here again on Chrome webworker memory usage.

This will bring us down a bunch.

Idea is: If we're adding a file in the background that's only a dependency, we don't keep the it in the parser heap, or the file contents hashtable.

We don't call any other functions on that file unless it gets hh_add_file'd later.

In another diff, I'm going to remove the file contents hashtable all together.

Reviewed By: @pikatchu

Differential Revision: D1567415