Make APC's StoreValue even smaller
Summary: Primed APC keys always have a zero expiration time, and non-primed
keys never need to use the lock.  Document that the lock is guaranteed to be
zero when unlocked so we can union it with expiry, and reduce the number of
bits we allow for expiry.  Also, rename 'expiry'.

Reviewed By: @edwinsmith

Differential Revision: D1567742